Datalink file movement
VS placed the datalink.cs file as a copy inside properties, relocated and made changes to adjust.
