+++ GameData/KCS/Parts/SmallScanner/smallScanner.cfg
@@ -79,6 +79,6 @@ PART
    	MODULE
   	{
 		name = ModuleObjectTracking
-		BaseDetectionRange = 500
+		baseDetectionRange = 500
   	}
 }
